+++ documents/src/commonMain/kotlin/io/github/nomemmurrakh/documents/Storage.kt
@@ -1,9 +1,0 @@
-package io.github.nomemmurrakh.documents
-
-internal interface Storage {
-    fun getBytes(key: String): ByteArray?
-    fun putBytes(key: String, value: ByteArray)
-    fun remove(key: String)
-    fun contains(key: String): Boolean
-    fun keys(prefix: String): List<String>
-}